functions/requirements.txt
azure-functions
azure-storage-blob
requests
lxml
pandas
beautifulsoup4
html5lib